% Generated by build script
\name{gdelt_articles}
\alias{gdelt_articles}
\title{Search GDELT for news articles}
\usage{
gdelt_articles()
}
\arguments{
\item{query}{Search query string (e.g., "climate change", "Ukraine")}
\item{maxrecords}{Maximum number of articles to return. Default 10, max 250.}
\item{timespan}{Timespan to search (e.g., "15min", "1h", "1d", "7d"). Default "7d".}
}
\value{tibble: url, title, seendate, socialimage, domain, language, sourcecountry}
\description{Search GDELT for news articles}
