% Generated by build script
\name{gdelt_timeline}
\alias{gdelt_timeline}
\title{Get GDELT timeline data for a query}
\usage{
gdelt_timeline()
}
\arguments{
\item{query}{Search query string}
\item{timespan}{Timespan (e.g., "7d", "30d", "1y"). Default "30d".}
}
\value{tibble: date, value, series}
\description{Get GDELT timeline data for a query}
